+++ packages/opensrp-reducer-factory/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.9 (2020-11-02)
+
+**Note:** Version bump only for package @opensrp/reducer-factory
+
+## 0.0.8 (2020-11-02)
+
+**Note:** Version bump only for package @opensrp/reducer-factory
+
+## [0.0.7](https://github.com/OpenSRP/web/compare/@opensrp/reducer-factory@0.0.6...@opensrp/reducer-factory@0.0.7) (2020-10-29)
+
+**Note:** Version bump only for package @opensrp/reducer-factory
+
+## 0.0.6 (2020-10-28)
+
+**Note:** Version bump only for package @opensrp/reducer-factory
+
+## 0.0.5 (2020-10-27)
+
+**Note:** Version bump only for package @opensrp/reducer-factory
+
+## 0.0.4 (2020-10-27)
+
+**Note:** Version bump only for package @opensrp/reducer-factory
+
 ## [0.0.3](https://github.com/opensrp/opensrp-web/compare/@opensrp/reducer-factory@0.0.1...@opensrp/reducer-factory@0.0.3) (2020-06-24)
 
 **Note:** Version bump only for package @opensrp/reducer-factory